chore(export): remove CRA, move super input files to src
