+++ packages/super-input/package.json
@@ -39,10 +39,7 @@
     "web-vitals": "^1.0.1"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
+    "build": "tsc",
     "build:doctree": "ts-node --skip-project ./scripts/genDocTree.ts"
   },
   "eslintConfig": {
