--- packages/super-input/src/SuperInput/index.tsx
@@ -1,7 +1,0 @@
-import SiBool from './SiBool'
-import SiTemplate from './SiTemplate'
-import SiExpression from './SiExpression'
-
-import './index.scss'
-
-export { SiBool, SiTemplate, SiExpression }
