+++ packages/super-input/src/colors.scss
