+++ packages/super-input/src/extensions/autocomplete/InfoCard.scss
