+++ packages/super-input/src/extensions/hoverInspect/InspectCard.tsx
